# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a Golang Application For Backend Candidate Test",
        "title": "Trading Application API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/users/order": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "ordering a product and calculate commission",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User ordering a product"
                ],
                "summary": "ordering a product and calculate commission",
                "parameters": [
                    {
                        "description": "Order product detail",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.OrderRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Order completed",
                        "schema": {
                            "$ref": "#/definitions/handlers.OrderResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "handlers.OrderRequest": {
            "type": "object",
            "properties": {
                "product_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "handlers.OrderResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "order_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "total_cost": {
                    "type": "number"
                }
            }
        },
...
    }
}